fix: Separate out components for simplicity
